libcusmm/autotune: add utilities file
- separate common utility variables/functions from hardware-specific
constraints
- complete hardware constraints
