+++ src/acc/libsmm_acc/libcusmm/kernels/cusmm_P100.py
@@ -1,20 +1,12 @@
-# hardware-dependent limits and sizes
-# for P100 
-# source: 
+# hardware-dependent constraints (GPU P100) 
+# source: CUDA occupancy calculator: 
+# http://developer.download.nvidia.com/compute/cuda/CUDA_Occupancy_calculator.xls
+
 maxTHREADSperBLOCK = 1024
 maxBLOCKSperSM = 32
 maxTHREADSperSM = 2048
 SMEMperBLOCK = 49152 # bytes
 SMEMperSM = 65536 # bytes
 warp_size = 32 # threads
-
-# sizeof
-sizeof_int = 4
-sizeof_double = 8
-
-def round_up_to_multiple(x, step):
-    if x % step == 0:
-        return x
-    else:
-        return x + step - x % step
+nSM = 56
